x86/smpboot.c: TXT AP bringup
In Intel TXT, all APs are in MONITOR/MWAIT loop on entry. Every AP uses
the same address for monitoring, so they are started simultaneously.

With this patch, every AP goes through assembly part, and only when in
start_secondary() in C they re-enter MONITOR/MWAIT iff they are not the
AP that was asked to boot. The same address is reused for simplicity,
and on next wakeup call APs don't have to go through assembly part
again (GDT, paging, stack setting).

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>